+++ packages/go/src/index.ts
@@ -28,13 +28,11 @@ import {
   glob,
   download,
   Lambda,
-  FileBlob,
   getWriteableDirectory,
   shouldServe,
   debug,
   cloneEnv,
   getProvidedRuntime,
-  getLambdaOptionsFromFunction,
 } from '@vercel/build-utils';
 
 const TMP = tmpdir();
@@ -47,35 +45,14 @@ import {
   OUT_EXTENSION,
 } from './go-helpers';
 
+import { GO_CANDIDATE_ENTRYPOINTS, detectGoEntrypoint } from './entrypoint';
+
+import {
+  buildStandaloneServer,
+  startStandaloneDevServer,
+} from './standalone-server';
+
 export { shouldServe };
-
-// Candidate entrypoints for Go runtime framework preset (in priority order)
-export const GO_CANDIDATE_ENTRYPOINTS = [
-  'main.go',
-  'cmd/api/main.go',
-  'cmd/server/main.go',
-];
-
-export async function detectGoEntrypoint(
-  workPath: string,
-  configuredEntrypoint: string
-): Promise<string | null> {
-  // If the configured entrypoint exists, use it
-  if (await pathExists(join(workPath, configuredEntrypoint))) {
-    debug(`Using configured Go entrypoint: ${configuredEntrypoint}`);
-    return configuredEntrypoint;
-  }
-
-  // Search candidate locations
-  for (const candidate of GO_CANDIDATE_ENTRYPOINTS) {
-    if (await pathExists(join(workPath, candidate))) {
-      debug(`Detected Go entrypoint: ${candidate}`);
-      return candidate;
-    }
-  }
-
-  return null;
-}
 
 // in order to allow the user to have `main.go`,
 // we need our `main.go` to be called something else
@@ -534,143 +511,6 @@ async function buildHandlerAsPackageMain({
   }
 }
 
-/**
- * Build a standalone Go HTTP server (runtime framework preset mode).
- * This builds a bootstrap wrapper that handles the Vercel IPC protocol
- * and proxies requests to the user's Go server.
- *
- * Architecture:
- * - User's server is built as 'user-server' binary
- * - Bootstrap wrapper is built as 'executable' (main entrypoint)
- * - Bootstrap handles IPC protocol and reverse proxies to user's server
- */
-async function buildStandaloneServer({
-  files,
-  entrypoint,
-  config,
-  workPath,
-  meta = {},
-}: BuildOptions): Promise<{ output: Lambda }> {
-  debug(`Building standalone Go server: ${entrypoint}`);
-
-  await download(files, workPath, meta);
-
-  // Get lambda options from config (memory, maxDuration, regions, architecture)
-  const lambdaOptions = await getLambdaOptionsFromFunction({
-    sourceFile: entrypoint,
-    config,
-  });
-
-  const architecture = lambdaOptions?.architecture || 'x86_64';
-
-  // Cross-compile for Linux with appropriate architecture
-  const env = cloneEnv(process.env, meta.env, {
-    GOARCH: architecture === 'arm64' ? 'arm64' : 'amd64',
-    GOOS: 'linux',
-    CGO_ENABLED: '0',
-  });
-
-  const { goModPath } = await findGoModPath(workPath, workPath);
-  const modulePath = goModPath ? dirname(goModPath) : workPath;
-
-  const go = await createGo({
-    modulePath,
-    opts: { cwd: workPath, env },
-    workPath,
-  });
-
-  const outDir = await getWriteableDirectory();
-  const userServerPath = join(outDir, 'user-server');
-  const bootstrapPath = join(outDir, 'executable');
-
-  // Determine build target based on entrypoint location
-  // - main.go at root: build '.'
-  // - cmd/api/main.go: build './cmd/api'
-  const buildTarget =
-    entrypoint === 'main.go' ? '.' : './' + dirname(entrypoint);
-
-  debug(
-    `Building user Go server (${architecture}): go build ${buildTarget} -> ${userServerPath}`
-  );
-
-  try {
-    await go.build(buildTarget, userServerPath);
-  } catch (err) {
-    console.error(`Failed to build standalone Go server: ${buildTarget}`);
-    throw err;
-  }
-
-  // Build the bootstrap wrapper that handles IPC protocol
-  const bootstrapSrc = join(__dirname, '../vercel-bootstrap.go');
-  debug(`Building bootstrap wrapper: ${bootstrapSrc} -> ${bootstrapPath}`);
-
-  try {
-    // Create a temporary directory for building the bootstrap
-    const bootstrapBuildDir = await getWriteableDirectory();
-    const bootstrapGoFile = join(bootstrapBuildDir, 'main.go');
-
-    // Copy bootstrap source to temp directory
-    await copy(bootstrapSrc, bootstrapGoFile);
-
-    // Initialize a minimal go.mod for the bootstrap
-    const bootstrapGoMod = join(bootstrapBuildDir, 'go.mod');
-    await writeFile(bootstrapGoMod, 'module vercel-bootstrap\n\ngo 1.21\n');
-
-    // Build bootstrap with same env (cross-compile settings)
-    const bootstrapGo = await createGo({
-      modulePath: bootstrapBuildDir,
-      opts: { cwd: bootstrapBuildDir, env },
-      workPath: bootstrapBuildDir,
-    });
-
-    await bootstrapGo.build('.', bootstrapPath);
-  } catch (err) {
-    console.error('Failed to build bootstrap wrapper');
-    throw err;
-  }
-
-  // Gather any additional files to include (user-specified via includeFiles)
-  // Note: Static files in public/static/ should be handled by the static builder,
-  // not bundled with the function. This follows the same pattern as Python.
-  const includedFiles: Files = {};
-  if (config && config.includeFiles) {
-    const patterns = Array.isArray(config.includeFiles)
-      ? config.includeFiles
-      : [config.includeFiles];
-    for (const pattern of patterns) {
-      const fsFiles = await glob(pattern, workPath);
-      for (const [assetName, asset] of Object.entries(fsFiles)) {
-        includedFiles[assetName] = asset;
-      }
-    }
-  }
-
-  // Read both binaries as FileBlob (embedded data)
-  // This ensures they persist regardless of temp file lifecycle
-  const [userServerData, bootstrapData] = await Promise.all([
-    readFile(userServerPath),
-    readFile(bootstrapPath),
-  ]);
-
-  const lambda = new Lambda({
-    ...lambdaOptions,
-    files: {
-      ...includedFiles,
-      // Bootstrap is the main entrypoint (handles IPC protocol)
-      executable: new FileBlob({ mode: 0o755, data: bootstrapData }),
-      // User's server is spawned by bootstrap
-      'user-server': new FileBlob({ mode: 0o755, data: userServerData }),
-    },
-    handler: 'executable',
-    runtime: 'executable',
-    supportsResponseStreaming: true,
-    architecture,
-    runtimeLanguage: 'go',
-  });
-
-  return { output: lambda };
-}
-
 async function renameHandlerFunction(fsPath: string, from: string, to: string) {
   let fileContents = await readFile(fsPath, 'utf8');
 
@@ -991,48 +831,6 @@ async function writeGoWork(
     .join('')})\n`;
   // console.log(contents);
   await writeFile(join(destDir, 'go.work'), contents, 'utf-8');
-}
-
-/**
- * Start a dev server for standalone Go server mode.
- * This runs `go run <target>` directly with the PORT environment variable.
- */
-async function startStandaloneDevServer(
-  opts: StartDevServerOptions,
-  resolvedEntrypoint: string
-): Promise<StartDevServerResult> {
-  const { workPath, meta = {} } = opts;
-
-  // Use a random port in the ephemeral range
-  const port = Math.floor(Math.random() * (65535 - 49152) + 49152);
-
-  const env = cloneEnv(process.env, meta.env, {
-    PORT: String(port),
-  });
-
-  // Determine run target based on entrypoint location
-  // - main.go at root: go run .
-  // - cmd/api/main.go: go run ./cmd/api
-  const runTarget =
-    resolvedEntrypoint === 'main.go' ? '.' : './' + dirname(resolvedEntrypoint);
-
-  debug(
-    `Starting standalone Go dev server: go run ${runTarget} (port ${port})`
-  );
-
-  const child = spawn('go', ['run', runTarget], {
-    cwd: workPath,
-    env,
-    stdio: ['ignore', 'inherit', 'inherit'],
-  });
-
-  // Give the server time to start
-  await new Promise(resolve => setTimeout(resolve, 2000));
-
-  return {
-    port,
-    pid: child.pid!,
-  };
 }
 
 export async function startDevServer(
